+++ system/src/Grav/Framework/Form/Traits/FormTrait.php
@@ -337,9 +337,14 @@ trait FormTrait
     {
         if (null === $this->flash) {
             $grav = Grav::instance();
-            $id = $this->getFlashId();
+            $config = [
+                'session_id' => $this->getFlashId() ?? '',
+                'unique_id' => $this->getUniqueId(),
+                'form_name' => $this->getName()
+            ];
 
-            $this->flash = new FormFlash($id ?? '', $this->getUniqueId(), $this->getName());
+
+            $this->flash = new FormFlash($config);
             $this->flash->setUrl($grav['uri']->url)->setUser($grav['user'] ?? null);
         }
 
@@ -364,7 +369,12 @@ trait FormTrait
         $list = [];
         /** @var \SplFileInfo $file */
         foreach (new \FilesystemIterator($folder) as $file) {
-            $flash = new FormFlash($id, $file->getFilename(), $name);
+            $config = [
+                'session_id' => $id,
+                'unique_id' => $file->getFilename(),
+                'form_name' => $name
+            ];
+            $flash = new FormFlash($config);
             if ($flash->exists() && $flash->getFormName() === $name) {
                 $list[] = $flash;
             }
@@ -403,10 +413,11 @@ trait FormTrait
         /** @var Grav $grav */
         $grav = Grav::instance();
 
-        $user = $grav['user'] ?? null;
-        if (isset($user)) {
-            $rememberState = $this->getBlueprint()->get('form/remember_state');
-            if ($rememberState === 'user') {
+        $rememberState = $this->getBlueprint()->get('form/remember_state');
+
+        if ($rememberState === 'user') {
+            $user = $grav['user'] ?? null;
+            if (isset($user)) {
                 return $user->username;
             }
         }